requirements.txt
# Core
flask>=2.0.0
gunicorn>=21.0.0

# Google Cloud
google-cloud-storage>=2.0.0

# SFTP
paramiko>=3.0.0
tenacity>=8.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
